+++ Assets/Scripts/CameraSmooth.cs
@@ -3,16 +3,9 @@ using UnityEngine;
 public class CameraSmooth : MonoBehaviour {
 
 	public Transform target;
-	public float smoothSpeed = 0.125f;
+	public float smoothSpeed = 0.1f;
 	public Vector3 PositionOffset;
-	public int FPS = 60;
-	float elapsed;
    Camera cam;
-	private void Start()
-	{
-		QualitySettings.vSyncCount = 1;
-		//Application.targetFrameRate = FPS;
-	}
 	void Update ()
 	{
 		transform.position = Vector3.Lerp(transform.position, target.position + PositionOffset, smoothSpeed);