+++ Assets/Scripts/Configurations/LightSettingManager.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class LightSettingManager : MonoBehaviour
+{
+    public List<GameObject> Lights;
+    public void OnStateChanges(bool isOn)
+    {
+        foreach (var light in Lights)
+        {
+            light.SetActive(isOn);
+        }
+    }
+}